Seems like it works on delete as well!!!

+++ lib/widgets/sweepstake_items.dart
@@ -5,6 +5,7 @@ import 'package:fancy_dialog/FancyAnimation.dart';
 import 'package:fancy_dialog/FancyGif.dart';
 import 'package:fancy_dialog/FancyTheme.dart';
 import 'package:fancy_dialog/fancy_dialog.dart';
+import 'package:flushbar/flushbar.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:halalbazaar/providers/great_places.dart';
@@ -20,6 +21,7 @@ class SweepstakeItems extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final productCount = Provider.of<GreatPlaces>(context, listen: false);
     return ListTile(
       contentPadding: EdgeInsets.all(4),
       leading: Container(
@@ -59,22 +61,32 @@ class SweepstakeItems extends StatelessWidget {
           IconButton(
         iconSize: 30,
         icon: Icon(Icons.delete),
-        onPressed: () {
-          showDialog(
-            context: context,
-            builder: (BuildContext context) => FancyDialog(
-              ok: 'Delete',
-              title: "We just want to confirm",
-              descreption: "Are you sure you want to delete this posting?",
-              animationType: FancyAnimation.BOTTOM_TOP,
-              theme: FancyTheme.FANCY,
-              gifPath: FancyGif.MOVE_FORWARD, //'./assets/walp.png',
-              okFun: () => {
-                Provider.of<GreatPlaces>(context, listen: false)
-                    .deleteProduct(id)
-              },
-            ),
-          );
+        onPressed: () async {
+          var descriptionCount = await productCount.getDescriptionValue();
+          if (descriptionCount > 0) {
+            Flushbar(
+              title: "Ohhh Shucks...",
+              message:
+                  "You can only have 5 active postings, please delete inactive or dated posts ",
+              duration: Duration(seconds: 5),
+            )..show(context);
+          } else {
+            showDialog(
+              context: context,
+              builder: (BuildContext context) => FancyDialog(
+                ok: 'Delete',
+                title: "We just want to confirm",
+                descreption: "Are you sure you want to delete this posting?",
+                animationType: FancyAnimation.BOTTOM_TOP,
+                theme: FancyTheme.FANCY,
+                gifPath: FancyGif.MOVE_FORWARD, //'./assets/walp.png',
+                okFun: () => {
+                  Provider.of<GreatPlaces>(context, listen: false)
+                      .deleteProduct(id)
+                },
+              ),
+            );
+          }
         },
         color: Theme.of(context).errorColor,
       ),
